feat: Add comprehensive tests for W3CResolver
This commit introduces a comprehensive test suite for the  contract, significantly increasing test coverage and ensuring the reliability of the DID resolution logic.

The new tests cover a wide range of scenarios, including:

- **Basic DID Resolution:** Verifies the correct resolution of basic DID documents.
- **Verification Methods:** Ensures that verification methods are correctly included in the resolved DID document.
- **Services:** Validates the inclusion of services in the DID document.
- **Method Expiration:** Tests the exclusion of expired verification methods.
- **Complex Scenarios:** Covers the resolution of complex DIDs with multiple verification methods and services.
- **Error Handling:** Includes tests for error conditions, such as resolving with empty DID input.
- **Coverage Gaps:** Addresses specific code paths that were not previously covered by tests.

This commit enhances the robustness of the  contract and provides a solid foundation for future development.

+++ script/Helper.sol
@@ -17,7 +17,7 @@ contract Helper {
     bytes memory trimmedBytes = new bytes(jsonBytes.length - 2);
 
     // Copy the characters, skipping the first and last ones
-    for (uint i = 1; i < jsonBytes.length - 1; i++) {
+    for (uint256 i = 1; i < jsonBytes.length - 1; i++) {
       trimmedBytes[i - 1] = jsonBytes[i];
     }
 
@@ -34,15 +34,15 @@ contract Helper {
       return new bytes(0);
     }
     bytes memory withoutZeros = new bytes(input.length);
-    uint length = 0;
-    for (uint i = 0; i < input.length; i++) {
+    uint256 length = 0;
+    for (uint256 i = 0; i < input.length; i++) {
       if (input[i] != 0x00) {
         withoutZeros[length] = input[i];
         length++;
       }
     }
     output = new bytes(length);
-    for (uint i = 0; i < length; i++) {
+    for (uint256 i = 0; i < length; i++) {
       output[i] = withoutZeros[i];
     }
     return output;
